feat(posts): add confirmation before deletion of post

+++ src/components/Dashboard/Post/Post.jsx
@@ -1,3 +1,4 @@
+import styles from "./Post.module.css";
 import { useParams, useNavigate } from "react-router";
 import { useState, useEffect } from "react";
 import { deletePost } from "../../../utils/deletePost";
@@ -25,9 +26,14 @@ export function Post() {
   if (!post) return <h1>Post not found! 404</h1>;
 
   async function handleDelete() {
-    const deleted = await deletePost(postId);
-    if (deleted) navigate("/");
-    else alert("Failed to delete post.");
+    const isConfirmed = window.confirm(
+      "Are you sure you want to delete this post? This action cannot be undone.",
+    );
+    if (isConfirmed) {
+      const deleted = await deletePost(postId);
+      if (deleted) navigate("/");
+      else alert("Failed to delete post.");
+    }
   }
 
   return (
@@ -35,7 +41,9 @@ export function Post() {
       <h3>{post.title}</h3>
       <div dangerouslySetInnerHTML={{ __html: post.content }} />
       <button onClick={() => navigate(`/posts/${post.id}/edit`)}>Edit</button>
-      <button onClick={handleDelete}>Delete</button>
+      <button onClick={handleDelete} className={styles.delete}>
+        Delete
+      </button>
       <Comments postId={post.id} />
     </article>
   );
